api/requirements.txt
alembic==1.16.5
celery==5.5.3
fastapi==0.116.1
httpx==0.28.1
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
prometheus-client==0.22.1
psycopg[binary]==3.2.10
pydantic==2.11.9
pydantic-settings==2.11.0
python-multipart==0.0.20
python-dotenv==1.1.1
redis==6.4.0
SQLAlchemy==2.0.43
uvicorn[standard]==0.35.0
